add exploreViewModel + remove logic from exploreView
