+++ PokemonSimpleApp/FeatureExplore/PokemonExploreView/PokemonExploreView.swift
@@ -10,32 +10,31 @@ import SwiftUI
 
 struct PokemonExploreView: View {
     
-    let getPokemonListUseCase: GetPokemonListUseCase = GetPokemonListUseCase(pokemonRepository: ExploreRepository.shared)
-    let limit: Int = 20
+    @StateObject var vm = PokemonListViewModel()
     private let networkMonitor = NetworkMonitor.shared
-    
-    @State private var pokemonList: [PokemonEntity] = []
-    @State private var offset: Int = 0
-    
-    @Environment(\.managedObjectContext) var context
     @FetchRequest(sortDescriptors: []) private var results: FetchedResults<Pokemon>
     
     var body: some View {
         NavigationStack {
             VStack {
                 List {
-                    if results.isEmpty || !results.isEmpty && !pokemonList.isEmpty {
-                        ForEach(pokemonList, id: \.self) { pokemon in
-                            NavigationLink(destination: PokemonDetailView(id: pokemon.id)) {
-                                PokemonListView(pokemon: pokemon)
-                                    .onAppear(perform: {
-                                        if pokemonList.last == pokemon {
-                                            loadMore()
+                    //                    if results.isEmpty || !results.isEmpty && !vm.pokemonList.isEmpty {
+                    ForEach(vm.pokemonList, id: \.self) { pokemon in
+                        NavigationLink(destination: PokemonDetailView(id: pokemon.id)) {
+                            PokemonListView(pokemon: pokemon)
+                                .task {
+                                    do {
+                                        if vm.pokemonList.last == pokemon {
+                                            try await vm.loadMore()
                                         }
-                                    })
-                            }
+                                        
+                                    } catch {
+                                        print("get pokeon list error")
+                                    }
+                                }
                         }
                     }
+                    //                    }
                 }
                 .navigationTitle("Explore Pokemons 🐉")
                 
@@ -43,19 +42,10 @@ struct PokemonExploreView: View {
             }
         }
         .task {
-            loadMore()
-        }
-    }
-    
-    private func loadMore() {
-        Task {
             do {
-                let newPokemonList = try await getPokemonListUseCase.execute(context: context, limit: limit, offset: offset)
-                pokemonList += newPokemonList
-                
-                offset += newPokemonList.count
+                try await vm.loadMore()
             } catch {
-                print("Error occurred: \(error)")
+                print("get pokeon list error")
             }
         }
     }
